pyproject.toml
[project]
name = "mahjong-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "mahjong>=1.2.0",
    "dspy-ai>=2.0.0",
]
